+++ src/lightning_habana/pytorch/strategies/deepspeed.py
@@ -32,13 +32,14 @@ from torch.optim import Optimizer
 
 if module_available("lightning"):
     from lightning.fabric.plugins import ClusterEnvironment
+    from lightning.fabric.strategies import _StrategyRegistry
     from lightning.fabric.utilities.optimizer import _optimizers_to_device
     from lightning.fabric.utilities.seed import reset_seed
     from lightning.fabric.utilities.types import _PATH, LRScheduler, ReduceLROnPlateau
     from lightning.pytorch import LightningModule, Trainer
     from lightning.pytorch.accelerators import Accelerator
     from lightning.pytorch.core.optimizer import _init_optimizers_and_lr_schedulers
-    from lightning.pytorch.overrides.base import _LightningModuleWrapperBase, _LightningPrecisionModuleWrapperBase
+    from lightning.pytorch.overrides.base import _LightningPrecisionModuleWrapperBase
     from lightning.pytorch.plugins.precision import PrecisionPlugin
     from lightning.pytorch.strategies.utils import _fp_to_half
     from lightning.pytorch.trainer.states import TrainerFn
@@ -46,16 +47,17 @@ if module_available("lightning"):
     from lightning.pytorch.utilities.exceptions import MisconfigurationException
     from lightning.pytorch.utilities.model_helpers import is_overridden
     from lightning.pytorch.utilities.rank_zero import WarningCache, rank_zero_info, rank_zero_only, rank_zero_warn
-    from lightning.pytorch.utilities.types import STEP_OUTPUT, LRSchedulerConfig
+    from lightning.pytorch.utilities.types import LRSchedulerConfig
 elif module_available("pytorch_lightning"):
     from lightning_fabric.plugins import ClusterEnvironment
+    from lightning_fabric.strategies import _StrategyRegistry
     from lightning_fabric.utilities.optimizer import _optimizers_to_device
     from lightning_fabric.utilities.seed import reset_seed
     from lightning_fabric.utilities.types import _PATH, LRScheduler, ReduceLROnPlateau
     from pytorch_lightning.core.optimizer import _init_optimizers_and_lr_schedulers
     from pytorch_lightning.accelerators import Accelerator
     from pytorch_lightning import LightningModule, Trainer
-    from pytorch_lightning.overrides.base import _LightningModuleWrapperBase, _LightningPrecisionModuleWrapperBase
+    from pytorch_lightning.overrides.base import _LightningPrecisionModuleWrapperBase
     from pytorch_lightning.plugins.precision import PrecisionPlugin
     from pytorch_lightning.strategies.utils import _fp_to_half
     from pytorch_lightning.trainer.states import TrainerFn
@@ -63,7 +65,7 @@ elif module_available("pytorch_lightning"):
     from pytorch_lightning.utilities.exceptions import MisconfigurationException
     from pytorch_lightning.utilities.model_helpers import is_overridden
     from pytorch_lightning.utilities.rank_zero import WarningCache, rank_zero_info, rank_zero_only, rank_zero_warn
-    from pytorch_lightning.utilities.types import STEP_OUTPUT, LRSchedulerConfig
+    from pytorch_lightning.utilities.types import LRSchedulerConfig
 
 from lightning_habana.pytorch.accelerator import HPUAccelerator
 from lightning_habana.pytorch.strategies.parallel import HPUParallelStrategy
@@ -445,12 +447,10 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
             )
 
         assert isinstance(self.model, (LightningModule, _LightningPrecisionModuleWrapperBase))
-        model = _LightningModuleWrapperBase(forward_module=self.model)
-
         if self.lightning_module.trainer and self.lightning_module.trainer.training:
-            self._initialize_deepspeed_train(model)
+            self._initialize_deepspeed_train(self.model)
         else:
-            self._initialize_deepspeed_inference(model)
+            self._initialize_deepspeed_inference(self.model)
 
     def _init_optimizers(self) -> Tuple[Optimizer, Optional[LRSchedulerConfig]]:
         assert self.lightning_module is not None
@@ -538,6 +538,7 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
                 profile=checkpoint_config.get("profile"),
             )
 
+    # TBD - this is still experimental and not complete
     def _initialize_deepspeed_inference(self, model: Module) -> None:
         import deepspeed
 
@@ -547,6 +548,8 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
         inference_config = {"train_micro_batch_size_per_gpu": 1}
         if "fp16" in self.config:
             inference_config.update({"fp16": self.config["fp16"]})
+        if "bf16" in self.config:
+            inference_config.update({"bf16": self.config["bf16"]})
         if self.zero_stage_3:
             inference_config.update(
                 {
@@ -865,7 +868,7 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
         pass
 
     @classmethod
-    def register_strategies(cls, strategy_registry: Dict) -> None:
+    def register_strategies(cls, strategy_registry: _StrategyRegistry) -> None:
         strategy_registry.register("hpu_deepspeed", cls, description="Default DeepSpeed Strategy")
         strategy_registry.register(
             "hpu_deepspeed_stage_1", cls, description="DeepSpeed with ZeRO Stage 1 enabled", stage=1
@@ -889,33 +892,7 @@ class HPUDeepSpeedStrategy(HPUParallelStrategy):
             offload_optimizer=True,
             offload_parameters=True,
         )
-        strategy_registry.register(
-            "hpu_deepspeed_stage_3_offload_nvme",
-            cls,
-            description="DeepSpeed ZeRO Stage 3 and NVMe Offload",
-            stage=3,
-            offload_optimizer=True,
-            offload_parameters=True,
-            remote_device="nvme",
-            offload_params_device="nvme",
-            offload_optimizer_device="nvme",
-        )
 
     def batch_to_device(self, batch: Any, device: Optional[torch.device] = None, dataloader_idx: int = 0) -> Any:
         batch = apply_to_collection(batch, Tensor, function=_fp_to_half, precision=self.precision_plugin.precision)
         return super().batch_to_device(batch, device, dataloader_idx)
-
-    def validation_step(self, *args: Any, **kwargs: Any) -> Optional[STEP_OUTPUT]:
-        assert self.model is not None
-        with self.precision_plugin.val_step_context():
-            return self.model(*args, **kwargs)
-
-    def test_step(self, *args: Any, **kwargs: Any) -> Optional[STEP_OUTPUT]:
-        assert self.model is not None
-        with self.precision_plugin.test_step_context():
-            return self.model(*args, **kwargs)
-
-    def predict_step(self, *args: Any, **kwargs: Any) -> STEP_OUTPUT:
-        assert self.model is not None
-        with self.precision_plugin.predict_step_context():
-            return self.model(*args, **kwargs)