+++ src/lightning_habana/pytorch/strategies/single.py
@@ -21,7 +21,6 @@ if module_available("lightning"):
     from lightning.fabric.utilities.types import _DEVICE
     from lightning.pytorch import LightningModule, Trainer
     from lightning.pytorch.accelerators import Accelerator
-    from lightning.pytorch.plugins.io.hpu_plugin import HPUCheckpointIO
     from lightning.pytorch.plugins.io.wrapper import _WrappingCheckpointIO
     from lightning.pytorch.plugins.precision import PrecisionPlugin
     from lightning.pytorch.strategies.single_device import SingleDeviceStrategy
@@ -30,7 +29,6 @@ elif module_available("pytorch_lightning"):
     from lightning_fabric.utilities.types import _DEVICE
     from pytorch_lightning import LightningModule, Trainer
     from pytorch_lightning.accelerators import Accelerator
-    from pytorch_lightning.plugins.io.hpu_plugin import HPUCheckpointIO
     from pytorch_lightning.plugins.io.wrapper import _WrappingCheckpointIO
     from pytorch_lightning.plugins.precision import PrecisionPlugin
     from pytorch_lightning.strategies.single_device import SingleDeviceStrategy
@@ -40,6 +38,7 @@ else:
 from torch.nn import Module
 from torch.optim.optimizer import Optimizer
 
+from lightning_habana.pytorch.plugins.io_plugin import HPUCheckpointIO
 from lightning_habana.utils.imports import _HABANA_FRAMEWORK_AVAILABLE
 
 if _HABANA_FRAMEWORK_AVAILABLE: