+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UnReleased] - 2023-MM-DD
+
+### Added
+-
+
+### Changed
+-
+
+### Fixed
+- Fixed wrong imports & update deep speed strategy ([#44](https://github.com/Lightning-AI/lightning-Habana/pull/44))
+
+### Removed
+-
+
+
+
 ## [1.0.0] - 2023-06-14
 
 ### Added